CAT model sub-plugins
Introduced sub-plugins to enable custom CAT logic in the adaptive quiz plugin.
Several interfaces were defined to allow for injectable logic. Sub-plugins
are expected to implement the interfaces in order to inject their custom
logic.
The custom logic relates to identifying what question should be administered
next or whether the attempt must be stopped with the reason provided.