CAT algorithm class: order of saving and calculating parameters.
The parameters based on the previous question answer are now calculated before determining the next difficulty level. This is
another step on taking some calculations logic out of the item administration class to make the logic replaceable.

The measure property is removed from the CAT algorithm class and converted to a method's parameter instead.

Made use of the existing value object for difficulty logit, added small extension to it and updated the PHPDocs.

Some unit tests were fixed, as the known state requirements have changed.