Question answer evaluation: move evaluation out of question answer processing.
The reason is to decouple the question answer evaluation from the processing of question answer. This will allow for further
refactoring to decouple processing of question answer from question administration.

Removed the option when a question wasn't answered from the evaluation, as this option is anyway always considered as an error.