Process question answer right after submission and redirect.
Moved saving of attempt state and results of calculation out of the item administration evaluation into the attempt script.
The script now redirects after processing the question answer.

Last attempted difficulty is now calculated inside the item administration evaluation class based on the last question answered
instead of taking it as a parameter.