Changes in item administration interface.
The purpose is to give the alternative implementations of the item administration interface more flexibility on assessing
questions answers. Instead of taking in a pre-defined question answer evaluation object, the item administration's method
takes in a slot number of the previous question. By using it, the item administration implementation ca obtain all the required
data to process question answer.

Converted the question answer evaluation service to a dependency of the default item administration evaluation.

Updated the example sub-plugin accordingly.