User own attempt report: respect custom CAT models
In case a custom CAT model is set for an adaptive quiz activity, the user will not see the measure column in their own attempts
listed. To support the change, refactored the related table class, made 'Show ability measure' setting in mod_form disappear once
a custom CAT model is chosen, and added forcibly defaulting this setting to 0 when creating/updating an adaptive quiz instance.