User attempts report: moved the report from the view page
In order to be able to generate a link to the report page depending on what CAT model is used, the attempts report for teachers
was moved to another page. The view page provides a link to the report, which in future updates will be able to get overridden
to point to the attempts report provided by the sub-plugin being used.