Delete attempt: allow sub-plugins to run their code
Wrapped the code which deletes an attempt into API to call.
Added callback for sub-plugins to run their code when an attempt is deleted. Also, if a sub-plugin
has its own attempts report and wants to enable attempts deletion through it, it can redirect
to a certain endpoint to delete an attempt. Added example implementation for the 'helloworld'
sub-plugin.