Fix deprecation in attempt class
Looks like the property was meant to be set onto $this->adpqattempt
instead. As the code has been working without this property set
onto $this->adpqattempt, and the 'timemodified' property was not
updated when setting the 'uniqueid' property, it's better to just
remove the dynamic assignment and not make it be
$this->adpqattempt->timemodified. Updating the 'timemodified'
property for an attempt here may have not obvious effects in
other areas.