CAT process API
Wrapped CAT session into a class with two public methods - one for administering new item
and another for processing the previously submitted item.
The purpose is to enable covering session with tests and introducing internal changes
in a reliable way.
The commit contains a functional test (calculation_steps_test.php) to test each step of
the algorithm. Potentially this test should be converted to Behat test(s).
Modifies certain parts of the existing classes/scripts to stop their preventing from
wrapping the code into an API class. For example, abandon the __destruct() logic in
the fetchquestion class.