Fix algorithm bug when continuing attempt
When continuing attempt, the attempt page determined the current difficulty by using the catalgo class.
The class' method, get_current_diff_level(), determined the level in a wrong way. The incorrect level
then passed further as the attempt page's parameter and caused discrepancies down the CAT algorithm
resulting in wrong ability measure values.

The catalgo's method was eliminated. A new method was added to the attempt class which when continuing
the attempt gets the current level by querying questions data from the database.