Custom attempt feedback with editor
Introduces Moodle's editor to edit custom attempt feedback.
Apart from the editor itself, there's an ability to use
placeholders in the custom feedback, when the user wants
to include some parameters of the adaptive quiz instance
into the feedback's text. The set of available placeholders
is limited though.
The database schema accommodates the feature by introducing
the following fields: 'attemptfeedbackenable' - this now
controls whether the custom feedback feature is used for
the activity instance, 'showabilitymeasurefeedback' and
'showabilitymeasuresummary' control whether the ability
measure is displayed on the attempt finish page and
the attempts summary page respectively. The existing
'showabilitymeasure' field becomes deprecated, but
isn't removed. Its value is the source of default values
for both new 'showabilitymeasurefeedback' and
'showabilitymeasuresummary' fields.
The new 'attemptfeedbackenable' field equals '-1' by default.
Such value means the instance is in transition period with
the new feedback feature - until a manager edits the instance
and overrides this field with either 0 or 1, the feedback
enabling will be treated as in the previous version - by emptiness
of the feedback's contents. The purpose of such setup is to
ensure a manager edits the feedback in the Moodle's editor first,
before putting 0 or 1 to the new 'attemptfeedbackenable' field.
This way we ensure the manager will see what they save in
the editor, as opposed to just putting the previous value
of the text to the editor's database field which could lead
to broken formatting of the text, as it was edited with no HTML
editor in the past.
The commit contains much refactoring of renderables and templates
which is another step away from the legacy code.
The commit alos refurbishes many plugin's tests. It appeared that
introducing of the new fields (see above), which are not nullable,
required editing of XML files serving as a source of test data
instead of generators. As using generators is the preferred
way to set up the test data, this commit revamps many existing
tests to start using the plugin's generators instead of the XML
files. This is another step away from the legacy.