Issue #45 grade: fix links in Gradebook
Depending on the user's role, proper links to adaptive quiz activity
should be displayed.
Apart from fixing the links themselves, the commit introduces
a persistent class for the attempt entity. The class is intended
to be a read model only, allowing for some convenient methods
to get information about attempt(s) - state, get a collection
of attempts by certain criteria, etc. The data manipulation
methods of the class are blocked in the initial implementation
to prevent any data manipulation in the class.